+++ routes/Wishlist.route.js
@@ -5,16 +5,20 @@ const wishapp = express.Router()
 const jwt = require("jsonwebtoken")
 
 wishapp.post("/add",async(req,res)=>{
-
+    const token = req.headers.authorization
 try{
- 
- const women = new WishListmodel(req.body)
- await women.save()
- //await WishListmodel.insertMany(req.body)
-  const data = await WishListmodel.find()
- res.status(200).send({"msg":"Item is added to WihsList","data":data})
-
-
+    if(token){
+        const decode =  jwt.verify(token,"batman")
+        if(decode){
+            const women = new WishListmodel(req.body)
+            await women.save()
+            //await WishListmodel.insertMany(req.body)
+            const data = await WishListmodel.find({"userID":decode.userID})
+            res.status(200).send({"msg":"Item is added to WihsList","data":data})
+        }
+        }else{
+            res.status(400).send({"msg":"login first"})
+        }
 }catch(err){
     res.status(400).send({"msg":"Item is already added to WishList"})
 }
@@ -24,7 +28,6 @@ try{
 
 wishapp.get("/get",async(req,res)=>{
     const token = req.headers.authorization
- 
 try{
     if(token){
        const decode =  jwt.verify(token,"batman")
@@ -43,10 +46,20 @@ try{
 
 wishapp.delete("/delete/:id",async(req,res)=>{
     const {id} = req.params
+    const token = req.headers.authorization
 try{
-  //await WishListmodel.deleteMany()  // if you want delete all data at once
- await WishListmodel.findByIdAndDelete({_id:id})
- res.status(200).send({"msg":"data is deleted"})
+    if(token){
+        const decode =  jwt.verify(token,"batman")
+        if(decode){
+          //await WishListmodel.deleteMany()  // if you want delete all data at once
+            await WishListmodel.findByIdAndDelete({_id:id})
+            const data = await WishListmodel.find({"userID":decode.userID})
+            res.status(200).send({"msg":"data is deleted","data":data})
+
+        }
+    }else{
+        res.status(400).send({"msg":"login first"}) 
+    } 
 }catch(err){
     res.status(400).send({"msg":err.message}) 
 }
